Remove premature rounding
This library does a lot of ROI locating on images. It also passes location values around. Pixel values were being rounded early on in multistep calculations, resulting in propagation of error. In particular CTP515 was "missing" its ROIs.

All premature rounding has been removed and now distance/location values are only rounded for locating pixels in final steps.

+++ src/alexandria/analyzers/ctp401.py
@@ -218,12 +218,12 @@ class CTP401Analyzer:
         if self.center is None:
             from alexandria.utils import find_center_edge_detection, compute_phantom_boundary, draw_boundary
 
-            def _unpack_center_result(value: Any) -> Tuple[int, int, Optional[float], Optional[float]]:
+            def _unpack_center_result(value: Any) -> Tuple[float, float, Optional[float], Optional[float]]:
                 if isinstance(value, (tuple, list)):
                     if len(value) >= 4:
-                        return int(value[0]), int(value[1]), value[2], value[3]
+                        return float(value[0]), float(value[1]), value[2], value[3]
                     if len(value) == 2:
-                        return int(value[0]), int(value[1]), None, None
+                        return float(value[0]), float(value[1]), None, None
                 raise ValueError(
                     "center_finder must return (row, col) or (row, col, diameter_y_px, diameter_x_px)"
                 )
